json builder clean up
