+++ LuckParser/Models/JsonModels/JsonMechanic.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LuckParser.Models
+{
+    public class JsonMechanic
+    {
+        public long Time;
+        public string Player;
+        public string Name;
+        public string Description;
+        public long Skill;
+    }
+}
