+++ LuckParser/Models/JsonModels/JsonPhase.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LuckParser.Models
+{
+    public class JsonPhase
+    {
+        public long Start;
+        public long End;
+        public string Name;
+        public int[] TargetIds;
+    }
+}
