+++ LuckParser/Models/JsonModels/JsonPlayer.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LuckParser.Models
+{
+    public class JsonPlayer
+    {
+        public string Character;
+        public string Account;
+        public uint Condition;
+        public uint Concentration;
+        public uint Healing;
+        public uint Toughness;
+        public int Group;
+        public string Profession;
+        public string[] Weapons;
+        public JsonDps DpsAll;
+        public JsonDps[] DpsBoss;
+        public JsonStatsAll StatsAll;
+        public JsonStats[] StatsBoss;
+        public JsonDefenses Defenses;
+        public JsonSupport Support;
+        public Dictionary<string, JsonBuffs> SelfBoons;
+        public Dictionary<string, JsonBuffs> GroupBoons;
+        public Dictionary<string, JsonBuffs> OffGroupBoons;
+        public Dictionary<string, JsonBuffs> SquadBoons;
+    }
+}
